requirements.txt
discord.py
fpl
python-dotenv